requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.3.0
numpy>=1.21.0